Moved functions in helpers module
