+++ napalm_base/clitools/cl_napalm_test.py
@@ -10,24 +10,23 @@ from __future__ import absolute_import
 
 # stdlib
 import sys
-import getpass
-import argparse
 
 import logging
 logger = logging.getLogger(__file__)
 
-# local modules
-from cl_napalm_configure import build_help
-from cl_napalm_configure import open_connection
-from cl_napalm_configure import configure_logging
+# import helpers
+from helpers import build_help
+from helpers import open_connection
+from helpers import configure_logging
 
 
 def main():
 
     args = build_help(connect_test=True)
-    configure_logging(args.debug)
+    configure_logging(logger, args.debug)
 
-    device = open_connection(args.vendor,
+    device = open_connection(logger,
+                             args.vendor,
                              args.hostname,
                              args.user,
                              args.password,
@@ -35,7 +34,11 @@ def main():
 
     print('Successfully connected to the device.')
 
+    logger.debug('Closing session...')
+
     device.close()
+
+    logger.debug('Connection closed!')
 
     sys.exit(0)
